creación de archivo para separar lógica de creación de matches
